+++ sandbox/index.ts
@@ -126,7 +126,7 @@ export class Sandbox {
     this.options = options;
     this.provider = service.provider;
     this.snapshotId = undefined;
-    this.version = '0.5';
+    this.version = '0.6';
 
     const accounts = this.getAccounts();