+++ scripts/build_blocklist.ts
@@ -4,6 +4,7 @@ import { network } from 'hardhat';
 import { Contract } from 'ethers';
 import { Event } from '@ethersproject/contracts';
 import { readNetworkConfig } from '../src/config';
+import { unwrap } from '../src/helpers';
 import { UNSRegistry__factory } from '../types/factories/contracts';
 import { CNSRegistry__factory } from '../types/factories/dot-crypto/contracts';
 
@@ -12,7 +13,7 @@ const UNSNetworkConfig = readNetworkConfig();
 async function main () {
   console.log('Network:', network.name);
 
-  const chainId: number = network.config.chainId!;
+  const chainId: number = unwrap(network.config, 'chainId');
   const config = UNSNetworkConfig.networks[chainId];
   if (!config) {
     throw new Error(`Config not found for network ${network.config.chainId}`);