+++ scripts/deploy_DotCoinBurner.ts
@@ -1,19 +1,18 @@
 import { network } from 'hardhat';
-import { mergeNetworkConfig, readNetworkConfig } from '../src/config';
+
+import { readNetworkConfig, mergeNetworkConfig } from '../src/config';
 import { Deployer } from '../src/deployer';
-import { unwrap } from '../src/helpers';
 
 async function main () {
   console.log('Network:', network.name);
 
-  const chainId: number = unwrap(network.config, 'chainId');
-  const config = readNetworkConfig()[chainId];
+  const config = readNetworkConfig().networks[network.config.chainId];
   if (!config) {
-    throw new Error(`Config not found for network ${chainId}`);
+    throw new Error(`Config not found for network ${network.config.chainId}`);
   }
 
   const deployer = await Deployer.create();
-  const deployConfig = await deployer.execute(['uns', 'uns_config_cns'], config);
+  const deployConfig = await deployer.execute(['dot_coin_burner'], config);
   mergeNetworkConfig(deployConfig);
 
   console.log('Deployed!');