+++ scripts/deploy_UNS.ts
@@ -1,13 +1,14 @@
 import { network } from 'hardhat';
 import { mergeNetworkConfig, readNetworkConfig } from '../src/config';
 import { Deployer } from '../src/deployer';
+import { unwrap } from '../src/helpers';
 
 const UNSNetworkConfig = readNetworkConfig();
 
 async function main () {
   console.log('Network:', network.name);
 
-  const chainId: number = network.config.chainId!;
+  const chainId: number = unwrap(network.config, 'chainId');
   const config = UNSNetworkConfig.networks[chainId];
   if (!config) {
     throw new Error(`Config not found for network ${chainId}`);