+++ scripts/upgrade_UNSRegistry.ts
@@ -1,11 +1,12 @@
 import { network } from 'hardhat';
 import { mergeNetworkConfig, readNetworkConfig } from '../src/config';
 import { Deployer } from '../src/deployer';
+import { unwrap } from '../src/helpers';
 
 async function main () {
   console.log('Network:', network.name);
 
-  const chainId: number = network.config.chainId!;
+  const chainId: number = unwrap(network.config, 'chainId');
   const config = readNetworkConfig()[chainId];
   if (!config) {
     throw new Error(`UNS config not found for network ${chainId}`);