+++ src/tasks.ts
@@ -453,7 +453,6 @@ const upgradeMintingManagerTask: Task = {
     const mintingManager = await upgrades.upgradeProxy(
       MintingManager.address,
       ctx.artifacts.MintingManager,
-      { unsafeAllowRenames: true },
     );
     await mintingManager.deployTransaction.wait();
 
@@ -699,16 +698,18 @@ const configureReconfigureTldL1Task = {
   ensureDependencies: (ctx: Deployer, config?: UnsNetworkConfig) => {
     config = merge(ctx.getDeployConfig(), config);
 
-    if (network.config.chainId !== 1 && network.config.chainId !== 5) {
+    const { chainId } = unwrap(network.config, 'chainId');
+
+    if (![1, 5].includes(chainId)) {
       throw new Error('Current network configuration does not support burning TLD');
     }
 
     const { MintingManager } = config.contracts || {};
     const dependencies = { MintingManager };
-    if (MintingManager.address) {
-      throw new Error(
-        `MintingManager contract not found for network ${network.config.chainId}`,
-      );
+    for (const [key, value] of Object.entries(dependencies)) {
+      if (!value || !value.address) {
+        throw new Error(`${key} contract not found for network ${chainId}`);
+      }
     }
 
     return dependencies;
@@ -746,16 +747,51 @@ const configureReconfigureTldL2Task: Task = {
   ensureDependencies: (ctx: Deployer, config?: UnsNetworkConfig) => {
     config = merge(ctx.getDeployConfig(), config);
 
-    if (network.config.chainId !== 137 && network.config.chainId !== 80001) {
+    const { chainId } = unwrap(network.config, 'chainId');
+    if (![137, 80001].includes(chainId)) {
       throw new Error('Current network configuration does not support moving TLD ownership');
     }
 
     const { MintingManager } = config.contracts || {};
     const dependencies = { MintingManager };
-    if (MintingManager.address) {
-      throw new Error(
-        `MintingManager contract not found for network ${network.config.chainId}`,
-      );
+    for (const [key, value] of Object.entries(dependencies)) {
+      if (!value || !value.address) {
+        throw new Error(`${key} contract not found for network ${chainId}`);
+      }
+    }
+
+    return dependencies;
+  },
+};
+
+const deployDotCoinBurnerTask = {
+  tags: ['dot_coin_burner', 'full'],
+  priority: 200,
+  run: async (ctx: Deployer, dependencies: DependenciesMap) => {
+    const { owner } = ctx.accounts;
+    const UNSRegistry = unwrap(dependencies, ArtifactName.MintingManager);
+
+    const dotCoinBurner = await ctx.artifacts.DotCoinBurner
+      .connect(owner)
+      .deploy(UNSRegistry.address);
+    await ctx.saveContractConfig(UnsContractName.DotCoinBurner, dotCoinBurner);
+    await dotCoinBurner.deployTransaction.wait();
+    await verify(ctx, dotCoinBurner.address, [UNSRegistry.address]);
+  },
+  ensureDependencies: (ctx: Deployer, config?: UnsNetworkConfig) => {
+    config = merge(ctx.getDeployConfig(), config);
+
+    const {
+      UNSRegistry,
+    } = config.contracts || {};
+    const dependencies = {
+      UNSRegistry,
+    };
+
+    for (const [key, value] of Object.entries(dependencies)) {
+      if (!value || !value.address) {
+        throw new Error(`${key} contract not found for network ${network.config.chainId}`);
+      }
     }
 
     return dependencies;
@@ -777,4 +813,5 @@ export const tasks: Task[] = [
   configureDotCoinTask,
   configureReconfigureTldL1Task,
   configureReconfigureTldL2Task,
+  deployDotCoinBurnerTask,
 ];